+++ web/keyApi.go
@@ -10,7 +10,6 @@ import (
 var (
 	keyAllApi    = regexp.MustCompile(`^\/api\/v1\/apikey[\/]*$`)
 	keySingleApi = regexp.MustCompile(`^\/api\/v1\/apikey\/(.*)$`)
-	keyListApi   = regexp.MustCompile(`^\/api\/v1\/apikeys[\/]*$`)
 )
 
 type keyAPI struct {
@@ -32,7 +31,7 @@ func (ka *keyAPI) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		ka.createApiKeyHandler(w, r)
 	case r.Method == http.MethodDelete && keySingleApi.MatchString(r.URL.Path):
 		ka.deleteApiKeyHandler(w, r)
-	case r.Method == http.MethodGet && keyListApi.MatchString(r.URL.Path):
+	case r.Method == http.MethodGet && keyAllApi.MatchString(r.URL.Path):
 		ka.listApiKeysHandler(w, r)
 	default:
 		notFoundApiError(w)
@@ -46,8 +45,7 @@ func (ka *keyAPI) listApiKeysHandler(w http.ResponseWriter, _ *http.Request) {
 		internalServerError(w)
 		return
 	}
-	response := map[string]interface{}{"apiKeys": apiKeys}
-	resJsonBytes, err := generateSuccessMessage(response)
+	resJsonBytes, err := generateSuccessMessage(apiKeys)
 	if err != nil {
 		internalServerError(w)
 		return
@@ -63,8 +61,7 @@ func (ka *keyAPI) createApiKeyHandler(w http.ResponseWriter, _ *http.Request) {
 		internalServerError(w)
 		return
 	}
-	response := map[string]string{"apiKey": apiKey}
-	resJsonBytes, err := generateSuccessMessage(response)
+	resJsonBytes, err := generateSuccessMessage(apiKey)
 	if err != nil {
 		internalServerError(w)
 		return
