+++ web/static/webcomponents/pages/apiKeys/apiKeys-home.js
@@ -1,9 +1,9 @@
-import { fromHTML } from "../../../utils.js";
+import { fromHTML } from "../../utils.js";
 
 const apikeysHomeTemplate = document.createElement("template");
 apikeysHomeTemplate.innerHTML = `
 <link rel="stylesheet" type="text/css" href="/static/reset.css">
-<link rel="stylesheet" type="text/css" href="/static/webcomponents/pages/apikeys/apikeys-home/apikeys-home.css">
+<link rel="stylesheet" type="text/css" href="/static/webcomponents/pages/apiKeys/apiKeys-home.css">
 `;
 
 class ApiKeysHome extends HTMLElement {
@@ -14,14 +14,43 @@ class ApiKeysHome extends HTMLElement {
     this.apiKeysData = [];
   }
 
+  renderApiKeys() {
+    const apiKeyCount = this.apiKeysData.length;
+    const apiKeysList = this.shadow.getElementById("api-keys-list");
+    apiKeysList.innerHTML = "";
+    this.apiKeysData.forEach((key, index) => {
+      const listItem = document.createElement("li");
+      listItem.innerHTML = `
+        <span>${apiKeyCount - index}. ${key.key}</span>
+        <button class="delete-button" data-key="${key.key}">Delete</button>
+      `;
+      apiKeysList.appendChild(listItem);
+    });
+    this.attachDeleteHandlers();
+  }
+
+  attachDeleteHandlers() {
+    console.log('Attaching delete handler')
+    const deleteButtons = this.shadow.querySelectorAll(".delete-button");
+    deleteButtons.forEach((button) => {
+      button.addEventListener("click", (event) => {
+        const deleteEvent = new CustomEvent("delete-api-key", {
+          detail: event.target.dataset.key,
+        });
+        // TODO: Change this from the shawdow DOM
+        this.shadow.dispatchEvent(deleteEvent);
+      });
+    });
+  }
+
   async fetchApiKeys() {
     try {
-      const response = await fetch("/api/v1/apikeys");
+      const response = await fetch("/api/v1/apikey");
       if (!response.ok) {
         throw new Error(`HTTP error! Status: ${response.status}`);
       }
       const data = await response.json();
-      this.apiKeysData = data.data.apiKeys.reverse(); // Reverse the order of the API keys
+      this.apiKeysData = data.data.reverse(); // Reverse the order of the API keys
       this.updateApiKeysComponent();
     } catch (error) {
       console.error("Error fetching API keys:", error);
@@ -55,11 +84,12 @@ class ApiKeysHome extends HTMLElement {
         throw new Error(`Failed to create API key: ${response.statusText}`);
       }
       const data = await response.json();
-      const newApiKey = { key: data.data.apiKey };
+      const newApiKey = { key: data.data };
       this.apiKeysData.unshift(newApiKey); // Add new key to the top of the list
       this.updateApiKeysComponent();
       console.log(`New API key created: ${newApiKey.key}`); // Log creation
       this.showSuccessMessage(newApiKey.key);
+      this.renderApiKeys()
     } catch (error) {
       console.error("Error creating API key:", error);
       this.showErrorMessage();
@@ -67,17 +97,17 @@ class ApiKeysHome extends HTMLElement {
   }
 
   updateApiKeysComponent() {
-    const apiKeysComponent = this.shadow.querySelector("api-keys");
-    if (apiKeysComponent) {
-      apiKeysComponent.apiKeys = this.apiKeysData;
-    }
+    this.renderApiKeys()
   }
 
   showSuccessMessage(apiKey) {
-    const apiKeysComponent = this.shadow.querySelector("api-keys");
-    if (apiKeysComponent) {
-      apiKeysComponent.displayNewApiKey(apiKey);
-    }
+    const modalElement = document.createElement("wc-modal");
+    modalElement.innerHTML = `
+      <div class="modal-content">
+          <p>New API Key: <span id="new-api-key">${apiKey}</span></p>
+      </div>
+    `;
+    this.shadow.appendChild(modalElement);
   }
 
   showErrorMessage() {
@@ -91,16 +121,15 @@ class ApiKeysHome extends HTMLElement {
     pageTitleElement.buttonFunction = this.createApiKey.bind(this);
     this.shadow.appendChild(pageTitleElement);
 
-    const apiKeysComponent = document.createElement("api-keys");
-    apiKeysComponent.addEventListener("create-api-key", () =>
-      this.createApiKey(),
-    );
-    apiKeysComponent.addEventListener("delete-api-key", (event) =>
-      this.deleteApiKey(event.detail),
-    );
-    this.shadow.appendChild(apiKeysComponent);
+    const keyContainerElement = fromHTML(`<div class="api-keys-container"><ul id="api-keys-list"></ul></div>`)
+    this.shadow.appendChild(keyContainerElement);
 
     this.fetchApiKeys();
+    // TODO: Change this from the shawdow DOM
+    this.shadow.addEventListener("delete-api-key", (event) => {
+      this.deleteApiKey(event.detail)
+      }
+    );
   }
 }
