+++ web/templates/baseApp.html
@@ -32,8 +32,7 @@
         <script defer type="module" src="/static/webcomponents/pages/topics/topics-home/topics-home.js"></script>
         <script defer type="module" src="/static/webcomponents/pages/topics/topic/topic.js"></script>
         <!-- API Keys -->
-        <script defer type="module" src="/static/webcomponents/pages/apikeys/apikeys-home/apikeys-home.js"></script>
-        <script defer type="module" src="/static/webcomponents/pages/apikeys/apikeys.js"></script>
+        <script defer type="module" src="/static/webcomponents/pages/apiKeys/apiKeys-home.js"></script>
 
     </head>
     <body class="appShellContainer">
